requirements-server.txt
# LangGraph Deep Research Server Requirements

# FastAPI and web server
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-groq>=0.2.0

# MCP (Model Context Protocol) integration
langchain-mcp-adapters>=0.1.0
mcp>=1.0.0

# Configuration and data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
typing-extensions>=4.8.0

# Web search and tools
tavily-python>=0.3.0
duckduckgo-search>=4.0.0

# HTTP requests and async
httpx>=0.25.0
aiohttp>=3.9.0
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Additional utilities for deep research
asyncio-throttle>=1.0.0

# Development and testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0 